LRAC-11227 set bounce as long just like the other fields

+++ modules/dxp/apps/osb/osb-asah/osb-asah-common/src/main/java/com/liferay/osb/asah/common/repository/impl/BQSessionRepositoryImpl.java
@@ -107,7 +107,7 @@ public class BQSessionRepositoryImpl
 						"averagesessionduration"
 					),
 					DSL.sum(
-						DSL.field("bounce", Integer.class)
+						DSL.field("bounce", Long.class)
 					).as(
 						"bounces"
 					),
